requirements.txt
joblib==0.11
pytz==2017.2
pyparsing==2.2.0
pyyaml==3.12
numpy==1.16
matplotlib==3.1.3
Pillow==6.2.0
scipy==1.2.0
scikit-image==0.15.0
scikit-learn==0.21
opencv-contrib-python==4.0.1.24
imutils==0.4.3
dlib==19.19.0
tensorflow-gpu==1.13.1
keras==2.1.3
keras-vis==0.4.1
tflearn==0.3.2